Remove flex rules from main container.
Overflown flex content does not scroll properly in Safari. Flex is not
really needed here, so we remove it and use standard block display
instead.

This patch also limits the height of the area with the chat messages and
the buttons to rougly equal the height of the viewport.
The benefit of limited height is that the user does not have to scroll
the screen down every time (s)he clicks a response button.