Fix avatar/message alignment when message is long.
When the message was long, the box with the message would be pushed down
below the avatar.

This patch fixes the problem by limiting the available width of both the
avatar and the message box.

This means that landscape avatar images will now be limited to a width
of 45px, but that is probably preferred anyway.